requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
groq==0.31.0
gTTS==2.5.4
python-multipart==0.0.20
